+++ test/unit/fuzz-tools.test.js
@@ -112,6 +112,7 @@ import * as toxV110 from '../../lib/tox-v110.js';
 import * as enviroV111 from '../../lib/enviro-v111.js';
 import * as critcareV112 from '../../lib/critcare-v112.js';
 import * as fluidrespV113 from '../../lib/fluidresp-v113.js';
+import * as pulmV114 from '../../lib/pulm-v114.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -168,6 +169,7 @@ const MODULES = {
   'enviro-v111.js': enviroV111,
   'critcare-v112.js': critcareV112,
   'fluidresp-v113.js': fluidrespV113,
+  'pulm-v114.js': pulmV114,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];